repo update

+++ src/global.js
@@ -1,6 +1,7 @@
 const { v1: uuidV1 } = require('uuid');
 const buttonHandles = {};
 const discord = require('discord.js');
+const zwc = require('./utils/zwc.js');
 
 module.exports = {
     client: null,
@@ -8,35 +9,36 @@ module.exports = {
         /**
          * 
          * @param {string} buttonId 
-         * @param {discord.Interaction} interaction 
+         * @param {discord.ButtonInteraction} interaction 
          */
         handle(buttonId, interaction) {
-            if (buttonHandles[buttonId]) {
-                if (buttonHandles[buttonId].owner != 0 &&
-                    buttonHandles[buttonId].owner == interaction.user.id) {
-                    interaction.deferUpdate();
-                    buttonHandles[buttonId](interaction);
-                    if (buttonHandles[buttonId].isOnce)
-                    {
-                        delete buttonHandles[buttonId];
-                    }
-                }
-                else {
-                    interaction.reply({
-                        content: '권한이 없습니다.',
-                        ephemeral: true
-                    });
+            if (!buttonHandles[buttonId]) {
+                this.errorMessage(interaction, `잘못된 요청입니다.`);
+            }
+
+            if (zwc.zwcToBytes(interaction.message.content).length <= 0) {
+                this.errorMessage(interaction, `잘못된 요청입니다.`);
+            }
+
+            if (buttonHandles[buttonId].owner == 0 ||
+                buttonHandles[buttonId].owner == interaction.user.id) {
+                interaction.deferUpdate();
+                buttonHandles[buttonId](interaction);
+                if (buttonHandles[buttonId].isOnce) {
+                    delete buttonHandles[buttonId];
                 }
             }
             else {
                 interaction.reply({
-                    content: '잘못된 요청입니다.',
+                    content: '권한이 없습니다',
                     ephemeral: true
                 });
-                interaction.message.edit({components: []});
             }
         },
-        add(button, handle, isOnce = true, owner = 0) {
+        add(label, style, handle, isOnce = true, owner = 0) {
+            const button = new discord.MessageButton();
+            button.setLabel(label);
+            button.setStyle(style);
             button.setCustomId(uuidV1());
             buttonHandles[button.customId] = handle;
             buttonHandles[button.customId].isOnce = isOnce;
@@ -48,10 +50,21 @@ module.exports = {
         }
     },
     message: {
+        /**
+         * 
+         * @param {discord.Message} message 
+         */
         delete(message) {
             setTimeout(() => {
-                message.delete();
+                if (message.deletable) message.delete();
             }, 300000);
         }
+    },
+    errorMessage(interaction, error) {
+        interaction.reply({
+            content: error,
+            ephemeral: true
+        });
+        interaction.message.edit({ components: [] });
     }
 }
